rename to UserSettingsPage2 and rollback UserSettingsPage
(cherry picked from commit 21a6aed9569baf06aaea2d1d4f2e90ca004ef1ce)
