+++ resources/i18n/el.json
@@ -410,7 +410,8 @@
     "Commit": "Δέσμευση",
     "EnterManually": "Εισάγετε χειροκίνητα",
     "Download": "Λήψη",
-    "Clone": "Κλώνος"
+    "Clone": "Κλώνος",
+    "Config": "Διαμόρφωση"
   },
   "agent": {
     "Endpoint": "Τελικό σημείο",
