+++ resources/i18n/it.json
@@ -410,7 +410,8 @@
     "Commit": "Impegno",
     "EnterManually": "Inserire manualmente",
     "Download": "Scarica",
-    "Clone": "Clone"
+    "Clone": "Clone",
+    "Config": "Configurazione"
   },
   "agent": {
     "Endpoint": "Endpoint",
