+++ resources/i18n/mn.json
@@ -410,7 +410,8 @@
     "Unset": "Тохируулаагүй",
     "OK": "БОЛЖ БАЙНА УУ",
     "Download": "Татаж авах",
-    "Clone": "Клон"
+    "Clone": "Клон",
+    "Config": "Тохиргоо"
   },
   "agent": {
     "Endpoint": "Төгсгөлийн цэг",
