+++ resources/i18n/vi.json
@@ -410,7 +410,8 @@
     "Commit": "Làm",
     "EnterManually": "Nhập thủ công",
     "Download": "Tải xuống",
-    "Clone": "Dòng vô tính"
+    "Clone": "Dòng vô tính",
+    "Config": "Cấu hình"
   },
   "agent": {
     "Endpoint": "Điểm cuối",
